Move --level option config to separate file
